Extract storage into class
